+++ UnamWebPanel/assets/php/templates.php
@@ -241,7 +241,7 @@ function unamtChart($charttype, $chartconfig){
 }
 
 function templateHashrateChart($label, $data) {
-    return unamtChart('hashrate', json_encode(['type'=>'bar', 'data'=>['datasets'=>[['label'=>$label, 'data'=>$data, 'fill'=>true]]], 'options'=>['responsive'=>true, 'scales'=>['x'=>['type'=>'time', 'max'=>date('Y-m-d H:i:00'), 'min'=>$data[0]['x'], 'time'=>['minUnit'=>'minute'], 'ticks'=>['color'=>'white']], 'y'=>['ticks'=>['color'=>'white', 'callback'=>''], 'min'=>0]]]]));
+    return unamtChart('hashrate', json_encode(['type'=>'bar', 'data'=>['datasets'=>[['label'=>$label, 'data'=>$data, 'fill'=>true]]], 'options'=>['responsive'=>true, 'scales'=>['x'=>['type'=>'time', 'max'=>date('Y-m-d H:i:00'), 'min'=>$data[0]['x'], 'time'=>['minUnit'=>'minute']], 'y'=>['ticks'=>['callback'=>''], 'min'=>0]]]]));
 }
 
 function templateDatatableTool($text, $checked, $options){